Move over `CustomParams` and `AdTargeting`
