+++ dotcom-rendering/index.d.ts
@@ -8,22 +8,6 @@ type SharePlatform =
 
 // shared type declarations
 
-type CustomParams = {
-	sens: 't' | 'f';
-	urlkw: string[];
-	[key: string]: string | string[] | number | number[] | boolean | boolean[];
-};
-
-type AdTargeting =
-	| {
-			adUnit: string;
-			customParams: CustomParams;
-			disableAds?: false;
-	  }
-	| {
-			disableAds: true;
-	  };
-
 interface SectionNielsenAPI {
 	name: string;
 	apiID: string;
